Porting Fixes for Issue 1770 on 3.2
* [CI] adding action for clang compiler
* [C] Ingnoring postgres warnings
* [dijkstra] compiling with clang
* [common] compiling with clang
* [astar] compiles with clang
* [ksp] compiles with clang
* [tsp] compiles with clang
* [alpha_shape] compiles with clang
* [tspEuclidean] compiles with clang
* [witPoints] compiles with clang
* [bellman_ford] compiles with clang
* [chinese] compiles with clang
* [transitiveClosure] compiles with clang
* [breadthFirstSearch] compiles with clang
* [components] compiles with clang
* [maxflow] pgr_maxflow.hpp compiles with clang
* [maxflow] compiles with clang
* [lineGraph] compiles with clang
* [bdAstar] compiles with clang
* [bdDijkstra] compiles with clang
* [trsp] compiles with clang
* [pickDeliver] compiles with clang
* [dijkstra][docqueries] giving a new correct alternate result
* Removing a test of a function that does not exist on pgRouting
* removing internal query tests that were created when developing contraction and are not used
* [tsp][docqueries] Modification of rand() gives new results
* roll back changes that affect ksp
* [C++] moving pgr_messages.cpp to cpp_common
* roll back changes that affect trsp
* src/common/basePath_SSEC.cpp compiles with clang
* src/trsp/pgr_trspHandler.cpp compiles with clang
* [ci][pgtap] adding testing after clang compilation
* [doc] adding NEWS & Release notes
* Also updating compilation_time and compiler & hash on pgr_full_version
* [build] rolling back build-extension-update-files1.pl
* [C++] adding some minimal changes from the comments of the PR
* Fixing merge conflicts

+++ src/lineGraph/lineGraphFull_driver.cpp
@@ -45,9 +45,7 @@ void get_turn_penalty_postgres_result(
         std::vector< Line_graph_full_rt > edge_result,
         Line_graph_full_rt **return_tuples,
         size_t &sequence) {
-    (*return_tuples) = pgr_alloc(
-            static_cast<int>(edge_result.size()),
-            (*return_tuples));
+    (*return_tuples) = pgr_alloc(edge_result.size(), (*return_tuples));
 
     for (const auto &edge : edge_result) {
         (*return_tuples)[sequence] =
